# Fix and optimize queries that need min or max block number