--- a/apps/block_scout_web/lib/block_scout_web/controllers/api/rpc/logs_controller.ex
+++ b/apps/block_scout_web/lib/block_scout_web/controllers/api/rpc/logs_controller.ex
@@ -183,7 +183,7 @@ defmodule BlockScoutWeb.API.RPC.LogsController do
   defp to_block_number(params, param_key) do
     case params[param_key] do
       "latest" ->
-        Chain.max_block_number()
+        Chain.consensus_block_number(:max)
 
       _ ->
         to_integer(params, param_key)