--- a/apps/block_scout_web/lib/block_scout_web/controllers/block_transaction_controller.ex
+++ b/apps/block_scout_web/lib/block_scout_web/controllers/block_transaction_controller.ex
@@ -82,7 +82,7 @@ defmodule BlockScoutWeb.BlockTransactionController do
   defp block_above_tip?("0x" <> _), do: nil
 
   defp block_above_tip?(block_hash_or_number) when is_binary(block_hash_or_number) do
-    with {:ok, max_block_number} <- Chain.max_block_number() do
+    with {:ok, max_block_number} <- Chain.consensus_block_number(:max) do
       {block_number, _} = Integer.parse(block_hash_or_number)
       block_number > max_block_number
     else