--- a/apps/block_scout_web/lib/block_scout_web/controllers/transaction_internal_transaction_controller.ex
+++ b/apps/block_scout_web/lib/block_scout_web/controllers/transaction_internal_transaction_controller.ex
@@ -37,14 +37,12 @@ defmodule BlockScoutWeb.TransactionInternalTransactionController do
 
       {internal_transactions, next_page} = split_list_by_page(internal_transactions_plus_one)
 
-      max_block_number = max_block_number()
-
       render(
         conn,
         "index.html",
         exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
         internal_transactions: internal_transactions,
-        max_block_number: max_block_number,
+        block_height: Chain.block_height(),
         show_token_transfers: Chain.transaction_has_token_transfers?(hash),
         next_page_params: next_page_params(next_page, internal_transactions, params),
         transaction: transaction
@@ -61,13 +59,6 @@ defmodule BlockScoutWeb.TransactionInternalTransactionController do
         |> put_status(404)
         |> put_view(TransactionView)
         |> render("not_found.html", transaction_hash: hash_string)
-    end
-  end
-
-  defp max_block_number do
-    case Chain.consensus_block_number(:max) do
-      {:ok, number} -> number
-      {:error, :not_found} -> 0
     end
   end
 end