--- a/apps/block_scout_web/lib/block_scout_web/controllers/transaction_log_controller.ex
+++ b/apps/block_scout_web/lib/block_scout_web/controllers/transaction_log_controller.ex
@@ -61,7 +61,7 @@ defmodule BlockScoutWeb.TransactionLogController do
   end
 
   defp max_block_number do
-    case Chain.max_block_number() do
+    case Chain.consensus_block_number(:max) do
       {:ok, number} -> number
       {:error, :not_found} -> 0
     end