--- a/apps/block_scout_web/lib/block_scout_web/views/block_transaction_view.ex
+++ b/apps/block_scout_web/lib/block_scout_web/views/block_transaction_view.ex
@@ -3,16 +3,15 @@ defmodule BlockScoutWeb.BlockTransactionView do
 
   import BlockScoutWeb.Gettext, only: [gettext: 1]
 
-  def block_not_found_message(block_above_tip) do
-    case block_above_tip do
-      true ->
-        gettext("Easy Cowboy! This block does not exist yet!")
+  def block_not_found_message({:ok, true}) do
+    gettext("Easy Cowboy! This block does not exist yet!")
+  end
 
-      false ->
-        gettext("This block has not been processed yet.")
+  def block_not_found_message({:ok, false}) do
+    gettext("This block has not been processed yet.")
+  end
 
-      _ ->
-        gettext("Block not found, please try again later.")
-    end
+  def block_not_found_message({:error, :hash}) do
+    gettext("Block not found, please try again later.")
   end
 end