--- a/apps/block_scout_web/lib/block_scout_web/views/transaction_view.ex
+++ b/apps/block_scout_web/lib/block_scout_web/views/transaction_view.ex
@@ -31,8 +31,12 @@ defmodule BlockScoutWeb.TransactionView do
 
   def confirmations(%Transaction{block: block}, named_arguments) when is_list(named_arguments) do
     case block do
-      nil -> 0
-      _ -> block |> Chain.confirmations(named_arguments) |> Cldr.Number.to_string!(format: "#,###")
+      nil ->
+        0
+
+      %Block{consensus: true} ->
+        {:ok, confirmations} = Chain.confirmations(block, named_arguments)
+        Cldr.Number.to_string!(confirmations, format: "#,###")
     end
   end