--- a/apps/indexer/lib/indexer.ex
+++ b/apps/indexer/lib/indexer.ex
@@ -22,7 +22,7 @@ defmodule Indexer do
 
   """
   def max_block_number do
-    case Chain.max_block_number() do
+    case Chain.consensus_block_number(:max) do
       {:ok, number} -> number
       {:error, :not_found} -> 0
     end