… apps/block_scout_web/test/block_scout_web/controllers/api/rpc/address_controller_test.exs
… apps/block_scout_web/test/block_scout_web/controllers/api/rpc/logs_controller_test.exs
… apps/block_scout_web/test/block_scout_web/controllers/api/rpc/transaction_controller_test.exs
… apps/block_scout_web/test/block_scout_web/controllers/block_transaction_controller_test.exs
… apps/block_scout_web/test/block_scout_web/features/viewing_app_test.exs
… apps/block_scout_web/test/block_scout_web/features/viewing_blocks_test.exs
… apps/block_scout_web/test/block_scout_web/views/transaction_view_test.exs
… apps/explorer/lib/explorer/etherscan.ex
… apps/explorer/test/explorer/chain_test.exs
… apps/explorer/test/explorer/etherscan_test.exs
… apps/explorer/test/support/factory.ex