Addresses part of #1328

## Changelog

### Enhancements
* By calculating the indexed ratio in SQL, we only have 1 round trip to the database instead of 1 to get the max and 1 to get the min.
* The block height should be used when calculating confirmations while the max consensus block number when a block number is needed and you want to tell if the chain is empty or only the genesis block is available.  The block height is 0 if there are no blocks or if there is only the genesis block.
* `Floki` selectors are used to improve error messages when messages are unexpected in `BlockTransactionControllerTest`.

### Bug Fixes
* Ensure min and max block number is for consensus blocks.  Adding `WHERE consensus = true` both fixes the bug that `min_block_number` and `max_block_number` were including non-consensus blocks, but it also makes the query faster because there is an index for consensus blocks.

  Times from Eth Mainnet production:

  ```
  ether=> SELECT min(b0."number") FROM "blocks" AS b0;
   min
  -----
     0
  (1 row)

  Time: 414.218 ms
  ether=> EXPLAIN SELECT min(b0."number") FROM "blocks" AS b0;
                                            QUERY PLAN
  -----------------------------------------------------------------------------------------------
   Finalize Aggregate  (cost=223454.95..223454.96 rows=1 width=8)
     ->  Gather  (cost=223454.73..223454.95 rows=2 width=8)
           Workers Planned: 2
           ->  Partial Aggregate  (cost=222454.73..222454.74 rows=1 width=8)
                 ->  Parallel Seq Scan on blocks b0  (cost=0.00..215051.99 rows=2961099 width=8)
  (5 rows)
  
  Time: 1.040 ms
  ether=> EXPLAIN SELECT min(b0."number") FROM "blocks" AS b0 WHERE b0.consensus = true;
                                                           QUERY PLAN
  ----------------------------------------------------------------------------------------------------------------------------
   Result  (cost=0.45..0.46 rows=1 width=8)
     InitPlan 1 (returns $0)
       ->  Limit  (cost=0.43..0.45 rows=1 width=8)
             ->  Index Only Scan using one_consensus_block_at_height on blocks b0  (cost=0.43..158050.90 rows=7045520 width=8)
                   Index Cond: (number IS NOT NULL)
  (5 rows)

  Time: 3.350 ms
  ether=> SELECT min(b0."number") FROM "blocks" AS b0 WHERE b0.consensus = true;
   min
  -----
     0
  (1 row)

  Time: 1.059 ms
  ```
* Tests for `BlockTransactionController` mistakenly simulated transactions on non-consensus blocks by making the `transaction.block` `consensus` `false`, but this is unrealistic - the importing only sets the consensus block to `transaction.block` and non-consensus blocks show up as `transaction.transaction_forks` `transaction_fork.uncle`.  

  To prevent this from happening again, `Explorer.Factory.with_block` now checks the `block` has `consensus: true`.  With this setup fixed, the tests needed to be updated to show

  1.  *No* _transactions_ would be shown for non-consensus blocks (because there are only transactions forks)
  2. Whether the non-consensus blocks showed as "above tip" only had to do with the number being above the consensus number and not to do with it being non-consensus.